+++ src/lib/app-provider/hooks/index.ts
@@ -2,3 +2,4 @@ export * from "./useFabricateFee";
 export * from "./useSimulateFee";
 export * from "./useRestrictedInput";
 export * from "./useCelatoneApp";
+export * from "./useQueryCmds";