+++ src/lib/components/modal/select-contract/ContractListDetail.tsx
@@ -1,103 +1,93 @@
 import { Box, Flex } from "@chakra-ui/react";
 import { matchSorter } from "match-sorter";
-import type { Dispatch, SetStateAction } from "react";
-import { useState } from "react";
+import { useMemo, useState } from "react";
 
 import { TagSelection, TextInput } from "lib/components/forms";
 import { EmptyState } from "lib/components/state/EmptyState";
 import { ZeroState } from "lib/components/state/ZeroState";
-import { useContractStore, useUserKey } from "lib/hooks";
-import { ContractListReadOnlyTable } from "lib/pages/contracts/components/ContractListReadOnlyTable";
-import { ContractListTable } from "lib/pages/contracts/components/ContractListTable";
+import { ContractListReadOnlyTable } from "lib/pages/contract-list/components/ContractListReadOnlyTable";
+import { ContractListTable } from "lib/pages/contract-list/components/ContractListTable";
 import type { ContractInfo, ContractListInfo } from "lib/stores/contract";
-import type { Option } from "lib/types";
+import type { ContractAddr, Option } from "lib/types";
 
 interface FilteredListDetailProps {
-  search: string;
-  tagFilter: string[];
   contracts: ContractInfo[];
-  isReadOnly: boolean;
-  isContractRemovable?: Option;
-  onContractSelect?: (addr: string) => void;
-}
-
-interface ListDetailProps {
-  search: string;
-  setSearch: Dispatch<SetStateAction<string>>;
-  contractListInfo: ContractListInfo;
-  isReadOnly: boolean;
-  isContractRemovable?: Option;
-  onContractSelect?: (addr: string) => void;
+  isReadOnly?: boolean;
+  contractRemovalInfo?: Option;
+  onContractSelect?: (addr: ContractAddr) => void;
 }
 
 const FilteredListDetail = ({
-  search,
-  tagFilter,
   contracts,
   isReadOnly,
-  isContractRemovable,
+  contractRemovalInfo,
   onContractSelect = () => {},
 }: FilteredListDetailProps) => {
-  const filteredContracts = matchSorter(contracts, search, {
-    keys: ["name", "description", "label", "address"],
-    sorter: (sortedItem) => sortedItem,
-  }).filter((contract) =>
-    tagFilter.every((tag) => contract.tags?.includes(tag))
-  );
-  if (filteredContracts.length === 0)
+  if (contracts.length === 0)
     return (
       <EmptyState
         message="No contracts match found. 
         Make sure you are searching with contract address, name, or description."
       />
     );
-  if (!isReadOnly)
-    return (
-      <ContractListTable
-        contracts={filteredContracts}
-        isContractRemovable={isContractRemovable}
-      />
-    );
-  return (
+
+  return isReadOnly ? (
     <ContractListReadOnlyTable
-      contracts={filteredContracts}
+      contracts={contracts}
       onContractSelect={onContractSelect}
+    />
+  ) : (
+    <ContractListTable
+      contracts={contracts}
+      contractRemovalInfo={contractRemovalInfo}
     />
   );
 };
 
-export const ListDetail = ({
-  search,
-  setSearch,
+interface ContractListDetailProps {
+  contractListInfo: ContractListInfo;
+  isReadOnly?: boolean;
+  isInstantiatedByMe?: boolean;
+  onContractSelect?: (addr: ContractAddr) => void;
+}
+
+export const ContractListDetail = ({
   contractListInfo,
   isReadOnly,
-  isContractRemovable,
+  isInstantiatedByMe = false,
   onContractSelect,
-}: ListDetailProps) => {
-  const userKey = useUserKey();
-  const { getAllTags } = useContractStore();
-
+}: ContractListDetailProps) => {
+  const [searchKeyword, setSearchKeyword] = useState("");
   const [tagFilter, setTagFilter] = useState<string[]>([]);
 
+  const filteredContracts = useMemo(
+    () =>
+      matchSorter(contractListInfo.contracts, searchKeyword, {
+        keys: ["name", "description", "label", "address"],
+        sorter: (sortedItem) => sortedItem,
+      }).filter((contract) =>
+        tagFilter.every((tag) => contract.tags?.includes(tag))
+      ),
+    [contractListInfo.contracts, searchKeyword, tagFilter]
+  );
+
   return (
-    <Box minH="xs">
+    <Box minH="xs" pb="48px">
       <Box px={isReadOnly ? "0px" : "48px"}>
         <Flex gap={2} w="full" my={isReadOnly ? "24px" : "48px"}>
           <TextInput
             variant="floating"
-            value={search}
-            setInputState={setSearch}
+            value={searchKeyword}
+            setInputState={setSearchKeyword}
             placeholder="Search with contract address, name, or description"
-            size="lg"
+            size={!isReadOnly ? "lg" : "md"}
           />
           {!isReadOnly && (
             <TagSelection
-              options={getAllTags(userKey)}
               result={tagFilter}
               setResult={setTagFilter}
               placeholder="No tag selected"
               label="Filter by tag"
-              labelBgColor="background.main"
               boxWidth="400px"
               creatable={false}
             />
@@ -108,14 +98,17 @@ export const ListDetail = ({
         <ZeroState
           list={{ label: contractListInfo.name, value: contractListInfo.slug }}
           isReadOnly={isReadOnly}
+          isInstantiatedByMe={isInstantiatedByMe}
         />
       ) : (
         <FilteredListDetail
-          search={search}
-          tagFilter={tagFilter}
-          contracts={contractListInfo.contracts}
+          contracts={filteredContracts}
           isReadOnly={isReadOnly}
-          isContractRemovable={isContractRemovable}
+          contractRemovalInfo={
+            contractListInfo.isContractRemovable
+              ? { label: contractListInfo.name, value: contractListInfo.slug }
+              : undefined
+          }
           onContractSelect={onContractSelect}
         />
       )}