+++ src/lib/layout/Header.tsx
@@ -1,15 +1,14 @@
 import {
   Flex,
-  Heading,
   Menu,
   MenuButton,
   MenuList,
   Text,
   MenuItem,
   Icon,
-  Image,
 } from "@chakra-ui/react";
 import { useWallet } from "@cosmos-kit/react";
+import Image from "next/image";
 import Link from "next/link";
 import { FiChevronDown } from "react-icons/fi";
 
@@ -29,9 +28,12 @@ const Header = () => {
       px={6}
     >
       <Link href="/">
-        <Heading as="h4" variant="h4">
-          <Image src="/celatone-logo.svg" alt="Celatone" width="160px" />
-        </Heading>
+        <Image
+          src="/celatone-logo.svg"
+          alt="Celatone"
+          width={120}
+          height={24}
+        />
       </Link>
       <Flex gap={2}>
         <Menu>