+++ src/lib/pages/query/components/QueryArea.tsx
@@ -5,7 +5,7 @@ import { useQuery } from "@tanstack/react-query";
 import type { AxiosError } from "axios";
 import { useEffect, useState } from "react";
 
-import ContractCmdButton from "lib/components/ContractCmdButton";
+import { ContractCmdButton } from "lib/components/ContractCmdButton";
 import { CopyButton } from "lib/components/CopyButton";
 import JsonInput from "lib/components/json/JsonInput";
 import JsonReadOnly from "lib/components/json/JsonReadOnly";
@@ -96,8 +96,7 @@ export const QueryArea = ({
               <ContractCmdButton
                 key={`query-cmd-${cmd}`}
                 cmd={cmd}
-                msg={jsonPrettify(queryMsg)}
-                setMsg={setMsg}
+                onClickCmd={() => setMsg(jsonPrettify(queryMsg))}
               />
             ))}
           </ButtonGroup>