+++ src/lib/stores/contract.ts
@@ -25,13 +25,8 @@ interface ContractList {
   isContractRemovable: boolean;
 }
 
-export interface ContractListInfo {
+export interface ContractListInfo extends Omit<ContractList, "contracts"> {
   contracts: ContractInfo[];
-  name: string;
-  slug: string;
-  lastUpdated: Date;
-  isInfoEditable: boolean;
-  isContractRemovable: boolean;
 }
 
 export const cmpContractListInfo = (